+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Release Date: ?
  - added the request uri exception message in Peak\Bedrock\Controller\FrontController
  - rounded time before outputting string in Peak\Common\TimeExpression
  - output milliseconds string if time is between 0 and 1 second in Peak\Common\TimeExpression
+ - added method toClockString() in Peak\Common\TimeExpression
 
 VERSION 2.19.0
 --------------